Split method tables across files.
